+++ src/process_handling/linux.rs
@@ -1,4 +1,5 @@
 use crate::collect_coverage;
+use crate::config::types::Mode;
 use crate::errors::*;
 use crate::event_log::*;
 use crate::process_handling::execute_test;
@@ -57,7 +58,11 @@ pub fn get_test_coverage(
                 }
             }
             Ok(ForkResult::Child) => {
-                info!("Launching test");
+                let bin_type = match config.command {
+                    Mode::Test => "test",
+                    Mode::Build => "binary",
+                };
+                info!("Launching {}", bin_type);
                 execute_test(test, ignored, config)?;
                 Ok(None)
             }