+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## Unreleased
+### Added
+
+### Changed
+- Updated logging so for the build mode it says "launching binary" instead of
+"launching test"
+
+### Removed
+
 ## [0.18.0-alpha2] 2021-04-16
 ### Added
 - Check if user sets -Cdebuginfo and remove it #601